+++ badge/ui/browser.h
@@ -1,9 +1,9 @@
 #ifndef INCLUDED_BADGE_UI_BROWSER_H
 #define INCLUDED_BADGE_UI_BROWSER_H
 
-#include <stddef.h>
+#include <stdint.h>
 
 void badge_browse_textfile(char const *fname);
-void badge_browse_textfiles(char const *menufile, size_t *selected);
+void badge_browse_textfiles(char const *menufile, uint8_t *selected);
 
 #endif
